Refactor code to proper modules & rename accordingly
